refactor: define HttpMethod
