+++ src/test/java/webserver/http/RequestLineTest.java
@@ -3,11 +3,13 @@ package webserver.http;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import webserver.RequestLine;
+import request.HttpMethod;
+import request.RequestLine;
 
 import java.util.Collections;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 /**
  * Created by youngjae.havi on 2019-08-01
@@ -18,7 +20,7 @@ public class RequestLineTest {
     void basic_parsing() {
         RequestLine requestLine = RequestLine.parse("GET /users HTTP/1.1");
 
-        assertThat(requestLine.getMethod()).isEqualTo("GET");
+        assertThat(requestLine.getMethod()).isEqualTo(HttpMethod.GET);
         assertThat(requestLine.getPath()).isEqualTo("/users");
         assertThat(requestLine.getHttpVersion()).isEqualTo("HTTP/1.1");
     }
@@ -27,7 +29,7 @@ public class RequestLineTest {
     void query_string_parsing() {
         RequestLine requestLine = RequestLine.parse("GET /users?userId=javajigi&password=password&name=JaeSung HTTP/1.1");
 
-        assertThat(requestLine.getMethod()).isEqualTo("GET");
+        assertThat(requestLine.getMethod()).isEqualTo(HttpMethod.GET);
         assertThat(requestLine.getPath()).isEqualTo("/users");
         assertThat(requestLine.getQueryString()).hasSize(3);
         assertThat(requestLine.getQueryString()).containsKeys("userId", "password", "name");
@@ -38,7 +40,7 @@ public class RequestLineTest {
     void not_include_query_string_parsing() {
         RequestLine requestLine = RequestLine.parse("GET /users HTTP/1.1");
 
-        assertThat(requestLine.getMethod()).isEqualTo("GET");
+        assertThat(requestLine.getMethod()).isEqualTo(HttpMethod.GET);
         assertThat(requestLine.getPath()).isEqualTo("/users");
         assertThat(requestLine.getQueryString()).hasSize(0);
         assertThat(requestLine.getHttpVersion()).isEqualTo("HTTP/1.1");
@@ -50,5 +52,10 @@ public class RequestLineTest {
         RequestLine requestLine = RequestLine.parse("GET /users?userId=javajigi&password=password&name=JaeSung HTTP/1.1");
         assertThat(requestLine.getQueryString()).containsKeys(key);
         assertThat(requestLine.getQueryString()).containsValue(Collections.singletonList(value));
+    }
+
+    @Test
+    void illegal_method() {
+        assertThrows(IllegalArgumentException.class, () -> RequestLine.parse("GOT /users HTTP/1.1"));
     }
 }
